Added default zoom template + moved view to zoom folder
