+++ src/cartodb.js
@@ -79,7 +79,7 @@ cdb.geo.ui.Text = require('./geo/ui/text');
 cdb.geo.ui.Annotation = require('./geo/ui/annotation');
 cdb.geo.ui.Image = require('./geo/ui/image');
 cdb.geo.ui.Share = require('./geo/ui/share');
-cdb.geo.ui.Zoom = require('./geo/ui/zoom');
+cdb.geo.ui.Zoom = require('./geo/ui/zoom/zoom-view');
 
 // setup expected object structure here, to avoid circular references
 _.extend(cdb.geo.ui, require('./geo/ui/legend-exports'));
