+++ src/geo/ui/mobile.js
@@ -7,7 +7,7 @@ var templates = require('cdb.templates');
 var sanitize = require('../../core/sanitize.js');
 var View = require('../../core/view');
 var Template = require('../../core/template');
-var Zoom = require('./zoom');
+var Zoom = require('./zoom/zoom-view');
 var TilesLoader = require('./tiles-loader');
 var Search = require('./search/search');
 var MobileLayer = require('./mobile-layer');
