+++ src/geo/ui/zoom/zoom-template.tpl
@@ -1,0 +1,5 @@
+<div class="CDB-Overlay">
+  <button class="CDB-Zoom-action CDB-Zoom-action--out js-zoomOut"></button>
+  <button class="CDB-Zoom-action CDB-Zoom-action--in js-zoomIn"></button>
+</div>
+<div class="CDB-Zoom-info js-zoomInfo">-</div>
